Adds mapping cassandra collection type to connect type
