+++ conf/cassandra-source-incr.properties
@@ -21,3 +21,4 @@ connect.cassandra.kcql=INSERT INTO orders-topic SELECT * FROM orders PK created 
 connect.cassandra.contact.points=localhost
 connect.cassandra.username=cassandra
 connect.cassandra.password=cassandra
+connect.cassandra.mapping.collection.to.json=true
