+++ conf/cassandra-source.properties
@@ -23,3 +23,4 @@ connect.cassandra.contact.points=localhost
 connect.cassandra.username=cassandra
 connect.cassandra.password=cassandra
 connect.cassandra.timestamp.type=timeuuid
+connect.cassandra.mapping.collection.to.json=true
